moved transactions interface to the common lib folder and removed kebab case file naming
